# Add basic lambda performance tests