@joe4dev 2023-05-03
👉  The performance issue on macOS is fixed in https://github.com/localstack/localstack/pull/8196