@joe4dev 2023-05-03
This PR adds e2e performance tests to reproduce the issue and easily evaluate Lambda performance in the future. The graphs need to be updated and then the PR should be ready for review after minor cleanup.