@joe4dev 2023-05-24
Updated performance results after the fix in https://github.com/localstack/localstack/pull/8196 and shortly before the 2.1 release (already indicated as 2.1.0 in the graph):
![localstack-lambda-performance](https://github.com/localstack/localstack/assets/8161292/bd7b6b8c-972c-47df-a38a-e9479ac2e8d8)

Hardware: Apple M1 Pro 16 GB non-binned 10/16 model
OS: macOS 13.3.1 (a) 

Graphs generated using: https://github.com/joe4dev/localstack-performance-testing
Dataset: https://github.com/joe4dev/localstack-performance-testing/tree/main/localstack-2.1.0

|    | label              |   percentile |   time_seconds |
|---:|:-------------------|----------:|---------------:|
|  0 | v1_docker_cold     |      0    |      1.73869   |
|  1 | v1_docker_cold     |      0.5  |      1.81394   |
|  2 | v1_docker_cold     |      0.95 |      1.91334   |
|  3 | v1_docker_cold     |      0.99 |      2.51415   |
|  4 | v1_docker_cold     |      1    |      2.68338   |
|  5 | v2.1.0_docker_cold |      0    |      1.11507   |
|  6 | v2.1.0_docker_cold |      0.5  |      1.19404   |
|  7 | v2.1.0_docker_cold |      0.95 |      1.27514   |
|  8 | v2.1.0_docker_cold |      0.99 |      1.29431   |
|  9 | v2.1.0_docker_cold |      1    |      1.37842   |
| 10 | v2.1.0_docker_warm |      0    |      0.0191527 |
| 11 | v2.1.0_docker_warm |      0.5  |      0.0302003 |
| 12 | v2.1.0_docker_warm |      0.95 |      0.0404974 |
| 13 | v2.1.0_docker_warm |      0.99 |      0.0640518 |
| 14 | v2.1.0_docker_warm |      1    |      0.0712791 |
| 15 | v2.1.0_host_cold   |      0    |      1.20881   |
| 16 | v2.1.0_host_cold   |      0.5  |      1.29218   |
| 17 | v2.1.0_host_cold   |      0.95 |      1.5028    |
| 18 | v2.1.0_host_cold   |      0.99 |      1.7876    |
| 19 | v2.1.0_host_cold   |      1    |      2.01963   |
| 20 | v2.1.0_host_warm   |      0    |      0.0187509 |
| 21 | v2.1.0_host_warm   |      0.5  |      0.0235497 |
| 22 | v2.1.0_host_warm   |      0.95 |      0.0302567 |
| 23 | v2.1.0_host_warm   |      0.99 |      0.0413241 |
| 24 | v2.1.0_host_warm   |      1    |      0.08435   |