@joe4dev 2023-06-06
Switch from `print` to `LOG.info` and rebased:

`print`

```
2023-06-06T14:31:59.819  INFO --- [   asgi_gw_6] localstack.request.aws     : AWS logs.PutLogEvents => 200
PASSED                                                                   [100%]
==================== EXECUTION TIME (s) for 100 repetitions ====================
0.019306000000142376 (min), 0.025516896001136047 (median), 0.05610130208104238 (p99), 0.08192070799850626 (max)

Process finished with exit code 0
```


`LOG.info`
```
2023-06-06T14:30:16.218  INFO --- [   asgi_gw_6] localstack.request.aws     : AWS lambda.Invoke => 200
2023-06-06T14:30:16.221  INFO --- [  MainThread] tests.integration.awslambda.test_lambda_performance : ==================== EXECUTION TIME (s) for 100 repetitions ====================
2023-06-06T14:30:16.221  INFO --- [  MainThread] tests.integration.awslambda.test_lambda_performance : 0.019451000000117347 (min), 0.02570416649996332 (median), 0.049755165420683624 (p99), 0.08308504200067546 (max)
PASSED                                                                   [100%]
Process finished with exit code 0
```